requirements.txt
aiohttp
beautifulsoup4
lxml
pandas
openpyxl
python-dotenv